Actualizando script para especificar opcionalmente el directorio de trabajo

+++ iniciar.sh
@@ -1,8 +1,21 @@
 #!/usr/bin/env bash
 
+set -euo pipefail
+
+DIRECTORY="$(pwd)"
+
+if (( $# > 0 )); then
+  DIRECTORY=$1
+fi
+
+if [[ ! -d "$DIRECTORY" ]]; then
+  echo "El directorio no existe. :("
+  exit 1
+fi
+
 docker run --name php \
            --rm -it \
            -p 80:80 \
-           -v "$(pwd):/opt/lampp/htdocs" \
+           -v "$DIRECTORY:/opt/lampp/htdocs" \
            tomsik68/xampp:7
